// src/app/components/tms/model/vehicle-requisition.model.ts
export class VehicleRequisitionModel {
        requisitionId: number;
        requisitionFor: number;
        requisitionForName: number;
        vehicleTypeId: number;
        buyerId: number;
        buyerName: string;
        Departments: any[];
        divisionId: number;
        divisionName: string;
        companyId: number;
        Employees: any[];
        employeeName: string;
        requisitionType: string;
        requiredFor: string;
        usedType: string;
        purpose: string;
        customerName: string;
        pickupLocation: string;
        dropLocation: string;
        startTime: Date;
        endTime: Date;
        noOfPerson: number;
        totalNoOfPerson: number;
        loadingCapacity: string;
        goodsWeight: string;
        goodsDetails: string;
        storeName: string;
        status: string;
        remarks: string;
        employee: string;
        reason: string;
        createBy: string;
        createDate: string;
        headOfficeId: string;
        branchOfficeId: string;
        requisitionSplit: string;
        mobileNo: string;
        registrationNumber: string;
        vehicleAssignMstId: number;
        assignmentId: number;
        createById: string;
        fromDate: Date;
        toDate: Date;
        buyerPersonName: string;
        buyerPersonMobileNo: string;
        participantName: string;
        supervisorRejectReason: string;
}